requirements.txt
pandas>=2.0
numpy>=1.26
matplotlib>=3.8
plotly>=5.22
chart-studio>=1.1
scikit-learn>=1.4
scipy>=1.12
openpyxl>=3.1
python-dotenv>=1.0
